Feature extraction and symmetry algorithms separated in own py files
